``` feat: Conquered a couple of algorithmic beasts! 🦁
SoberWalk.java:  Finally cracked the code for that tricky walk –  no more stumbling around in the dark (algorithmically speaking, of course!).  Optimized for elegance and efficiency, this solution's a beauty.

MajorityElement.java:  This one put up a good fight, but I managed to find that elusive majority element with a clever approach.  Feeling pretty smart after this one!  💪
```